pyproject.toml
[project]
name = "u2net"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.2.1",
    "pillow-heif>=0.22.0",
    "pyyaml>=6.0.2",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128" },
]
torchvision = [
    { index = "pytorch-cu128" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
